Move widgets, use only Pathlib
